change double into decimal

+++ BLL/IManager.cs
@@ -2,9 +2,9 @@
 {
     public interface IManager
     {
-        void AddChfByCardId(int cardId, double amountChf);
-        void AddChfByUsername(string username, double amountChf);
-        double GetChfbyUsername(string username);
-        double GetQuantityByUsername(string username);
+        void AddChfByCardId(int cardId, decimal amountChf);
+        void AddChfByUsername(string username, decimal amountChf);
+        decimal GetChfbyUsername(string username);
+        decimal GetQuantityByUsername(string username);
     }
 }
